--- src/main/java/janggi/Pieces.java
@@ -1,13 +1,0 @@
-package janggi;
-
-import janggi.piece.Piece;
-import java.util.List;
-
-public final class Pieces {
-
-    private final List<Piece> pieces;
-
-    public Pieces(final List<Piece> pieces) {
-        this.pieces = pieces;
-    }
-}
