+++ src/main/java/janggi/position/Column.java
@@ -1,8 +1,8 @@
-package janggi;
+package janggi.position;
 
 import java.util.Arrays;
 
-public enum Row {
+public enum Column {
 
     ZERO(0),
     ONE(1),
@@ -12,12 +12,11 @@ public enum Row {
     FIVE(5),
     SIX(6),
     SEVEN(7),
-    EIGHT(8),
-    NINE(9);
+    EIGHT(8);
 
     private final int value;
 
-    Row(final int value) {
+    Column(final int value) {
         this.value = value;
     }
 
@@ -25,16 +24,16 @@ public enum Row {
         return value;
     }
 
-    public Row add(final int dx) {
+    public Column add(final int dx) {
         return Arrays.stream(values())
                 .filter(newValue -> newValue.getValue() == value + dx)
                 .findAny()
                 .orElseThrow(() -> new IllegalStateException("[ERROR] 보드를 벗어난 값입니다."));
     }
 
-    public static Row of(final int value) {
+    public static Column of(final int value) {
         return Arrays.stream(values())
-                .filter(row -> row.getValue() == value)
+                .filter(column -> column.getValue() == value)
                 .findAny()
                 .orElseThrow(() -> new IllegalStateException("[ERROR] 보드를 벗어난 값입니다."));
     }
