+++ src/main/java/janggi/position/Row.java
@@ -1,8 +1,8 @@
-package janggi;
+package janggi.position;
 
 import java.util.Arrays;
 
-public enum Column {
+public enum Row {
 
     ZERO(0),
     ONE(1),
@@ -12,11 +12,12 @@ public enum Column {
     FIVE(5),
     SIX(6),
     SEVEN(7),
-    EIGHT(8);
+    EIGHT(8),
+    NINE(9);
 
     private final int value;
 
-    Column(final int value) {
+    Row(final int value) {
         this.value = value;
     }
 
@@ -24,16 +25,16 @@ public enum Column {
         return value;
     }
 
-    public Column add(final int dx) {
+    public Row add(final int dx) {
         return Arrays.stream(values())
                 .filter(newValue -> newValue.getValue() == value + dx)
                 .findAny()
                 .orElseThrow(() -> new IllegalStateException("[ERROR] 보드를 벗어난 값입니다."));
     }
 
-    public static Column of(final int value) {
+    public static Row of(final int value) {
         return Arrays.stream(values())
-                .filter(column -> column.getValue() == value)
+                .filter(row -> row.getValue() == value)
                 .findAny()
                 .orElseThrow(() -> new IllegalStateException("[ERROR] 보드를 벗어난 값입니다."));
     }
