+++ src/main/java/janggi/rule/MovingRule.java
@@ -1,4 +1,4 @@
-package janggi;
+package janggi.rule;
 
 import java.util.ArrayList;
 import java.util.List;
