+++ src/test/java/janggi/position/PositionTest.java
@@ -1,7 +1,8 @@
-package janggi;
+package janggi.position;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import janggi.rule.MoveVector;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
